refactor classes

+++ demo_rpg/include/demo_rpg/types.h
@@ -1,9 +1,8 @@
 #pragma once
 #include <cstdint>
-typedef std::uint16_t stattype;
-typedef std::uint16_t welltype;
-typedef std::uint16_t damagetype;
-typedef std::uint16_t itemcount;
-
-typedef std::uint64_t exptype;
-typedef std::uint16_t leveltype;
+typedef std::uint32_t stattype;
+typedef std::uint32_t welltype;
+typedef std::uint32_t damagetype;
+typedef std::uint32_t itemcount;
+typedef std::uint32_t exptype;
+typedef std::uint32_t leveltype;
